wip(npu): Task 7 FFN chain — plumbing works, numerics do not
ffn_chain.py runs phases P4+P5 in one dispatch (2 of 5 phases, 29.2 of the
layer's 38.0 MB). It builds, places and runs end to end with one operand fifo
and one result fifo shared between the phases and P4's results drained into the
buffer P5's fills read. It does NOT pass, and is marked as such in its docstring
so nothing builds on it.

Established, and reusable:

- fill/drain take offset, sizes, strides — the mechanism Task 7 needs and the
  plan never names. A phase's drain can write into a chosen slice of the buffer
  a later phase's fill reads, so P4's per-pair drains land directly in the
  activation halves of P5's broadcast objects. transfer_len alone emits
  sizes=[0,0,0,0] and is rejected; use sizes=[1,1,1,N], strides=[0,0,0,1].
- Phases must share one operand and one result fifo. Separate ones are a compile
  error, not a tight fit: 3 input DMA channels requested, 2 available, because
  the broadcast already takes one.
- Row assignment must match the JOIN order, not the core. A pair's join emits
  [core0's NROWS][core1's NROWS] per step, so a contiguous per-core block makes
  the pair's drain an interleaving of two distant ranges. row = pair*rpp +
  tile*2*NROWS + j*NROWS is what makes a pair's drain contiguous; this alone
  took the error from 1.38e-01 to 4.05e-02. DIM_ACCN must then cover a pair's
  row span, not a core's.

Not established: the values. SwiGLU relative error is p50 10.3% / max 51%
against a 5.86% exp2 floor, so it is a bug. The sorted multiset also mismatches,
so it is not pure ordering; 5505 of 8192 rows are exact and the rest scattered.
Candidates are the in-place RMSNorm prologue against a broadcast fifo shared
across phases, and the gate/up stash under a shared operand fifo. The isolation
is to run P4 alone with flm_asum_prepare and compare against ffn_alt.py, which
passes at 3.40e-04 with the same kernels and a private fifo.

Also switches flm_gemv_flush to emit bf16 (the layer output is the next layer's
residual stream, and every inter-phase value is bf16); down_verify.py updated
and still passes at 6.03e-04.

+++ kernels/npu/flm_gemv_flush.cc
@@ -14,16 +14,21 @@
 
 extern float g_acc_down[];
 
+// Emits **bf16**, not f32. This is the layer's output `x_out`, which is the
+// next layer's residual stream, and every inter-phase value in the fused layer
+// is bf16 — the broadcast object is bf16, so an f32 result would have to be
+// narrowed before the next phase could read it anyway. It also halves the
+// result traffic and lets one result object shape serve every phase.
 extern "C" __attribute__((noinline)) void
 flm_gemv_flush(const bfloat16 *restrict act_aux, const uint8 *restrict wtile,
-               float *restrict out) {
+               bfloat16 *restrict out) {
   float part[NROWS];
   flm_q4_1_tile(act_aux, wtile, part);
   const int base = tile_row_base(wtile);
   const int slot = base % DIM_ACCN;
   const bfloat16 *restrict aux = act_aux + K;
   for (int r = 0; r < NROWS; ++r) {
-    out[r] = g_acc_down[slot + r] + part[r] + float(aux[base + r]);
+    out[r] = bfloat16(g_acc_down[slot + r] + part[r] + float(aux[base + r]));
     g_acc_down[slot + r] = 0.0f;   // ready for the next token
   }
 }
